add clickhouse to github action
