+++ v03_pipeline/lib/misc/clickhouse_test.py
@@ -1,0 +1,10 @@
+import unittest
+
+from v03_pipeline.lib.misc.clickhouse import get_clickhouse_client
+
+
+class ClickhouseTest(unittest.TestCase):
+    def test_get_clickhouse_client(self):
+        client = get_clickhouse_client()
+        result = client.query('SELECT 1')
+        self.assertEqual(result.result_rows, [[1]])
