+++ v03_pipeline/lib/model/environment.py
@@ -30,7 +30,7 @@ CLICKHOUSE_SERVICE_HOSTNAME = os.environ.get(
     'seqr-clickhouse',
 )
 CLICKHOUSE_SERVICE_PORT = (int(os.environ.get('CLICKHOUSE_SERVICE_PORT', '9000')),)
-CLICKHOUSE_USERNAME = (os.environ.get('CLICKHOUSE_USERNAME', 'clickhouse'),)
+CLICKHOUSE_USER = (os.environ.get('CLICKHOUSE_USER', 'clickhouse'),)
 HAIL_BACKEND_SERVICE_HOSTNAME = os.environ.get(
     'HAIL_BACKEND_SERVICE_HOSTNAME',
     'hail-search',
@@ -58,7 +58,7 @@ class Env:
     CLICKHOUSE_PASSWORD: str = CLICKHOUSE_PASSWORD
     CLICKHOUSE_SERVICE_HOSTNAME: str = CLICKHOUSE_SERVICE_HOSTNAME
     CLICKHOUSE_SERVICE_PORT: int = CLICKHOUSE_SERVICE_PORT
-    CLICKHOUSE_USERNAME: str = CLICKHOUSE_USERNAME
+    CLICKHOUSE_USER: str = CLICKHOUSE_USER
     CLINGEN_ALLELE_REGISTRY_LOGIN: str | None = CLINGEN_ALLELE_REGISTRY_LOGIN
     CLINGEN_ALLELE_REGISTRY_PASSWORD: str | None = CLINGEN_ALLELE_REGISTRY_PASSWORD
     DEPLOYMENT_TYPE: Literal['dev', 'prod'] = DEPLOYMENT_TYPE
